+++ app.py
@@ -1,4 +1,4 @@
-from om_core import create_app
+from core import create_app
 
 # To do: This place will change later
 config = {
