+++ core/__init__.py
@@ -3,21 +3,29 @@ import os
 from flask import Flask
 from flask_cors import CORS
 from dotenv import load_dotenv
+from flask_sqlalchemy import SQLAlchemy
+
 
 from config import BaseConfig
 
 APP_ROOT = os.path.join(os.path.dirname(__file__))
 dotenv_path = os.path.join(APP_ROOT, '.env')
 load_dotenv(dotenv_path)
+database_file = "sqlite:///{}".format(os.path.join(APP_ROOT, "sampledb.db"))
+
+app = Flask(__name__)
+app.config["SQLALCHEMY_DATABASE_URI"] = database_file
+db = SQLAlchemy(app)
 
 def create_app(environment):
-    app = Flask(__name__)
 
     env = os.getenv("ENV")
 
     app.config.from_object(environment.get(env))
 
-    from om_core.api.user.controllers import user
+    from core.api.user.controller import user
+    from core.api.post.controller import post
+
 
     """ Cors settings will be here. We maybe use this endpoint later. """
     cors = CORS(app, resources={
@@ -31,5 +39,7 @@ def create_app(environment):
 
 
     app.register_blueprint(user, url_prefix='/api/users')
+    app.register_blueprint(post, url_prefix='/api/post')
+
 
     return app
