+++ core/api/user/model.py
@@ -1,0 +1,11 @@
+from core import db
+
+
+class User(db.Model):
+    """ Model for user management """
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(100))
+
+    def __init__(self, name):
+        self.name = name
