--- om_core/data/README.md
@@ -1,11 +1,0 @@
-# Models
-
-All data fields will store in the **models.py** file.
-
-## Current Model Classes
-
-These are the current model classes;
-
-**User**
-
-This model will work for user operations.
